Status system rework (#592)

+++ Assembly-CSharp/Global/SB2/SB2_MON_PARM.cs
@@ -61,7 +61,7 @@ public class SB2_MON_PARM
 
     public UInt16 Radius;
 
-    public UInt16 Geo;
+    public Int16 Geo;
 
     public UInt16[] Mot;
 
@@ -163,7 +163,7 @@ public class SB2_MON_PARM
     public Int32 WeaponAttachment;
 
     [Memoria.PatchableFieldAttribute]
-    public Byte[] TranceGlowingColor;
+    public Int32[] TranceGlowingColor;
 
     [Memoria.PatchableFieldAttribute]
     public UInt32 MaxDamageLimit;
